Changes "None" and "Remove" keywords
None and remove still remove the base effect if used as a replacement.
As a fallback, None now does nothing. It just prevents an error from
popping up.
As a fallback, Remove now removes the base effect, but ONLY IF it's a
fallback for a replacement.
Also added logging of fallbacks.